Refactor: Move getObjectKey utility to a separate file and update imports in uploadToAlgolia and uploadToS3
