+++ gazette-parse/uploadToAlgolia.ts
@@ -2,7 +2,8 @@ import { Parser } from "htmlparser2";
 const { PDFImage } = require("pdf-image");
 const Tesseract = require("tesseract.js");
 import algoliasearch from "algoliasearch";
-import { getObjectKey, parseMetadataCsv } from "./utils";
+import { parseMetadataCsv } from "./utils";
+import { getObjectKey } from "./utils/getObjectKey";
 const { PdfReader } = require("pdfreader");
 import * as fs from "fs";
 import path from "path";
