+++ gazette-parse/utils.ts
@@ -10,31 +10,6 @@ export type CsvMetadata = {
   title: string;
 };
 
-export type ObjectMetadata = {
-  notificationNumber: string;
-  fileName: string;
-  category: string;
-  subCategory: string;
-  year: number;
-};
-
-export const getObjectKey = ({
-  notificationNumber,
-  fileName,
-  year,
-  category,
-  subCategory,
-}: ObjectMetadata) => {
-  const fileNumber = notificationNumber
-    ? notificationNumber
-    : `${fileName.replace(".pdf", "").replace(".htm", "")}`;
-  const isPdfFile = fileName.includes(".pdf");
-  const objectKey = `${year}/${category}/${subCategory}/${fileNumber}.${
-    isPdfFile ? "pdf" : "htm"
-  }`;
-  return objectKey;
-};
-
 export const parseMetadataCsv = async (
   filePath: string
 ): Promise<CsvMetadata[]> => {
